03.1 de junio

+++ src/main/frontend/src/app/components/admin/alumnado/alumnado-lista/alumnado-lista.component.ts
@@ -9,6 +9,8 @@ import { MatDialog } from '@angular/material/dialog';
 import { AlumnadoAsignarClaseComponent } from '../alumnado-asignar-clase/alumnado-asignar-clase.component';
 import { NewClase } from '../../../../interfaces/newClase';
 import Swal from 'sweetalert2';
+import { ProfesoresService } from '../../../../services/profesores.service';
+import { Profesor } from '../../../../interfaces/profesor';
 
 @Component({
   selector: 'app-alumnado-lista',
@@ -70,7 +72,7 @@ import Swal from 'sweetalert2';
           <button class="mr-3" [routerLink]="['/admin/alumnado/editar', alumno.id]">
               <i class="fa-solid fa-pencil-alt"></i>
           </button>
-          <button (click)="deleteAlumno(alumno.id)">
+          <button (click)="deleteAlumno(alumno.id, alumno)">
               <i class="fa-solid fa-trash-alt"></i>
           </button>
         </td>
@@ -82,11 +84,12 @@ import Swal from 'sweetalert2';
 })
 export class AlumnadoListaComponent implements OnInit {
 
-  constructor(private alumnosService: AlumnosService, private claseService: ClasesService, private dialog: MatDialog) { }
+  constructor(private alumnosService: AlumnosService, private claseService: ClasesService, private dialog: MatDialog, private profesoresService: ProfesoresService) { }
 
   alumnosList: Alumno[] = [];
   nombreCurso: string = '';
   clases: Clase[] = [];
+  profesores: Profesor[] = [];
   clase: Clase | undefined;
 
   ngOnInit() {
@@ -107,6 +110,15 @@ export class AlumnadoListaComponent implements OnInit {
         console.error('Error al obtener la lista de clases:', error);
       }
     });
+
+    this.profesoresService.getAll().subscribe({
+      next: (profesores: Profesor[]) => {
+        this.profesores = profesores;
+      },
+      error: (error) => {
+        console.error('Error al obtener la lista de profesores:', error);
+      }
+    });
   }
 
   openAsignar(alumno: Alumno) {
@@ -119,22 +131,53 @@ export class AlumnadoListaComponent implements OnInit {
 
     dialog.afterClosed().subscribe((result: NewClase) => {
       if (result) {
-        this.claseService.create(result).subscribe({
-          next: () => {
-            Swal.fire({
-              title: "¡Hecho!",
-              text: "Clase asignada al alumno " + alumno.nombre + ".",
-              showConfirmButton: false,
-              timer: 1700,
-              icon: "success"
-            }).then(() => {
-              window.location.reload();
+
+        let clasesProfesor: Clase[] = [];
+        let claseOcupada: boolean = false;
+        let hayClases: boolean = false;
+
+        this.profesores.forEach(profesor => {
+          if (profesor.idInstrumento === alumno.idInstrumento) {
+            this.clases.forEach(clase => {
+              if (clase.idProfesor === profesor!.id) {
+                clasesProfesor.push(clase);
+                hayClases = true;
+              }
             });
-          },
-          error: (error) => {
-            console.error('Error al generar clase.');
           }
         });
+
+        if (hayClases) {
+          clasesProfesor.forEach(clase => {
+            if (clase.dia === result.dia && clase.hora === result.hora) {
+              Swal.fire({
+                title: "¡Error!",
+                text: "Ya hay una clase asignada a ese horario.",
+                icon: "error"
+              });
+              claseOcupada = true;
+            }
+          });
+        }
+
+        if (!claseOcupada) {
+          this.claseService.create(result).subscribe({
+            next: () => {
+              Swal.fire({
+                title: "¡Hecho!",
+                text: "Clase asignada al alumno " + alumno.nombre + " " + alumno.apellidos + ".",
+                showConfirmButton: false,
+                timer: 1700,
+                icon: "success"
+              }).then(() => {
+                window.location.reload();
+              });
+            },
+            error: (error) => {
+              console.error('Error al generar clase.');
+            }
+          });
+        }
       } else {
         console.log('El diálogo se cerró sin pasar datos');
       }
@@ -151,13 +194,35 @@ export class AlumnadoListaComponent implements OnInit {
     return false;
   }
 
-  deleteAlumno(id: number) {
-    this.alumnosService.delete(id).subscribe({
-      next: (response) => {
-        window.location.reload();
-      },
-      error: () => {
-        console.error('Error al eliminar el alumno con código ' + id);
+  deleteAlumno(id: number, alumno: Alumno) {
+    Swal.fire({
+      title: "¿Está seguro de borrar al alumno " + alumno.nombre + " " + alumno.apellidos + "?",
+      text: "No podrá revertir este cambio.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, borrar"
+    }).then((result) => {
+
+      if (result.isConfirmed) {
+
+        this.alumnosService.delete(id).subscribe({
+          next: () => { },
+          error: () => {
+            console.error('Error al eliminar el alumno con código ' + id);
+          }
+        });
+
+        Swal.fire({
+          title: "¡Hecho!",
+          text: "El alumno ya no pertenece a la escuela.",
+          showConfirmButton: false,
+          timer: 1700,
+          icon: "success"
+        }).then(() => {
+          window.location.reload();
+        });
       }
     });
   }
